fix: migrate lhs to audit-builder
